Move to test/e2e
