--- test/development/build-indicator/test/index.test.ts
@@ -1,87 +1,0 @@
-/* eslint-env jest */
-import { join } from 'path'
-import { retry } from 'next-test-utils'
-import { nextTestSetup } from 'e2e-utils'
-
-const installCheckVisible = (browser) => {
-  return browser.eval(`(function() {
-      window.checkInterval = setInterval(function() {
-      const root = document.querySelector('nextjs-portal').shadowRoot;
-      const indicator = root.querySelector('[data-next-mark]')
-      if(!indicator) return
-      window.showedBuilder = window.showedBuilder || (
-        indicator.getAttribute('data-next-mark-loading') === 'true'
-      )
-      if (window.showedBuilder) clearInterval(window.checkInterval)
-    }, 5)
-  })()`)
-}
-
-describe('Build Activity Indicator', () => {
-  describe('Invalid position config', () => {
-    const { next } = nextTestSetup({
-      files: join(__dirname, '..'),
-      skipStart: true,
-      startServerTimeout: 1000,
-      nextConfig: {
-        devIndicators: {
-          // Intentionally invalid position to test error
-          position: 'ttop-leff' as any,
-        },
-      },
-    })
-
-    it('should validate position config', async () => {
-      try {
-        await next.start()
-      } catch (err) {
-        // Ignore error
-      }
-      await retry(async () => {
-        expect(next.cliOutput).toInclude(
-          `Invalid "devIndicator.position" provided, expected one of top-left, top-right, bottom-left, bottom-right, received ttop-leff`
-        )
-      })
-    })
-  })
-
-  describe.each(['pages', 'app'])('Enabled - (%s)', (pagesOrApp) => {
-    const { next } = nextTestSetup({
-      files: join(__dirname, '..'),
-    })
-
-    ;(process.env.IS_TURBOPACK_TEST ? describe.skip : describe)(
-      'webpack only',
-      () => {
-        it('Shows the build indicator when a page is built during navigation', async () => {
-          const browser = await next.browser(
-            pagesOrApp === 'pages' ? '/' : '/app'
-          )
-          await installCheckVisible(browser)
-          await browser.elementByCss('#to-a').click()
-          await retry(async () => {
-            const wasVisible = await browser.eval('window.showedBuilder')
-            expect(wasVisible).toBe(true)
-          })
-        })
-      }
-    )
-
-    it('Shows build indicator when page is built from modifying', async () => {
-      const browser = await next.browser(
-        pagesOrApp === 'pages' ? '/b' : '/app/b'
-      )
-      await installCheckVisible(browser)
-      const pagePath =
-        pagesOrApp === 'pages' ? 'pages/b.js' : 'app/app/b/page.js'
-
-      await next.patchFile(pagePath, (content) => content.replace('b', 'c'))
-
-      await retry(async () => {
-        const wasVisible = await browser.eval('window.showedBuilder')
-
-        expect(wasVisible).toBe(true)
-      })
-    })
-  })
-})
