+++ test/e2e/build-indicator/pages/index.js
